+++ ChangeLog.rst
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Unreleased
 ----------
 
+Added
++++++
+
+* Add functions relating to exponentiation: exp, log, logn, log10, power
+
 v0.3 - 2018-03-31
 -----------------